gravier: parse_condition improve to support multiples condition
but im stupide and went the wrong way.
I check multiples relational perator, and skip logical one.
even worst, it seems perl doesn't define in which order relational
operator must be evaluated, so A > B > c, will be evalutabe like this.
if (A > (B > C)).

and gravier does it the other way.

anyway, it was some useless code, that should be change in next commit.

Signed-off-by: Matthias Gatto <uso.cosmo.ray@gmail.com>